+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
-No changes yet
+### Added
+- `Writer`, the writer monad
 
 ## [5.0.0] - 2019-09-23
 ### Changed